# rest/public/documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API REST de la plateforme MGLSI_NEWS",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/projet_blog_mglsi/rest",
            "description": "Cette url est la racine de toutes les endpoints à interroger."
        }
    ],
    "paths": {
        "/index.php?action=articles&type={type}": {
            "get": {
                "tags": [
                    "article"
                ],
                "operationId": "ArticleApiController::get",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Permet de spécifier le format de données (xml/json) que va retourner la requête",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nos articles disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "article"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "articles",
                                        "wrapped": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/index.php?action=article&id={id}&type={type}": {
            "get": {
                "tags": [
                    "article"
                ],
                "operationId": "PostController::get_by_id",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Permet de spécifier le format de données (xml/json) que va retourner la requête.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permet de spécifier l'id de la ressource.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "L'article en question.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/index.php?action=articlescategorie&categorie={categorie}&type={type}": {
            "get": {
                "tags": [
                    "article"
                ],
                "operationId": "PostController::get_by_cat",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Permet de spécifier le format de données (xml/json) que va retourner la requête.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categorie",
                        "in": "path",
                        "description": "Permet de spécifier la catégorie dont on veut récuperer les articles.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Les articles correspondant à la catégorie spécifier.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "article"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "articles",
                                        "wrapped": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/index.php?action=articlesByCategory&type={type}": {
            "get": {
                "tags": [
                    "article"
                ],
                "operationId": "PostController::get_all_cat",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Permet de spécifier le format de données (xml/json) que va retourner la requête.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nos articles disponibles regroupés par catégorie.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "article"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    },
                                    "xml": {
                                        "name": "articles",
                                        "wrapped": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "properties": {
                    "id": {
                        "description": "L'identifiant de l'article",
                        "type": "integer"
                    },
                    "titre": {
                        "description": "Titre de l'article",
                        "type": "string"
                    },
                    "contenu": {
                        "description": "Contenu de l'article",
                        "type": "string"
                    },
                    "dateCreation": {
                        "description": "Date de création de l'article",
                        "type": "string",
                        "format": "date-time"
                    },
                    "dateModification": {
                        "description": "Date de modification de l'article",
                        "type": "string",
                        "format": "date-time"
                    },
                    "categorie": {
                        "description": "Catégorie de l'article",
                        "$ref": "#/components/schemas/Categorie"
                    },
                    "auteur": {
                        "description": "Auteur de l'article",
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Categorie": {
                "properties": {
                    "id": {
                        "description": "L'identifiant de la catégorie",
                        "type": "integer"
                    },
                    "libelle": {
                        "description": "Libellé de la catégorie",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "L'identifiant de la l'utilisateur",
                        "type": "integer"
                    },
                    "nom": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "prenom": {
                        "description": "Prénom de l'utilisateur",
                        "type": "string"
                    },
                    "mail": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string"
                    },
                    "statut": {
                        "description": "Statut de l'utilisateur",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "article",
            "description": "Opérations et informations disponibles sur les articles .",
            "externalDocs": {
                "description": "Visitez le site du MGLSI_NEWS",
                "url": "http://localhost/projet_blog_mglsi"
            }
        }
    ]
}